+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added `truncate_ragged_lines` option to `read_csv`, `read_csv_batched`, and `scan_csv` methods
 - Added precompiled gem for Linux x86-64 MUSL
 - Changed `drop` method to ignore missing columns
+- Fixed error with `then` method
 
 ## 0.9.0 (2024-03-03)